Rename mod display -> forme
