+++ crates/kas-core/src/text/forme.rs
